refactor(develop): update development ux
