+++ src/components/index.tsx
@@ -1,0 +1,8 @@
+import * as Anchor from "./Anchor";
+export { Anchor };
+
+import * as Dendrogram from "./Dendrogram";
+export { Dendrogram };
+
+import * as Tree from "./Tree";
+export { Tree };
